Safari-611.1.10/LayoutTests/media/now-playing-status-without-media-expected.txt


NowPlaying status should not be updated when no media is active.


** Check state initially.
EXPECTED (state.haveEverRegisteredAsNowPlayingApplication == 'false') OK

** Suspend process.
RUN(internals.applicationWillBecomeInactive())
EXPECTED (state.haveEverRegisteredAsNowPlayingApplication == 'false') OK

** Activate process.
RUN(internals.applicationDidBecomeActive())
EXPECTED (state.haveEverRegisteredAsNowPlayingApplication == 'false') OK
END OF TEST
